[C]: add some conversions. Fix several warnings on Windows builds. Silence gtest.

+++ aeron-driver/src/main/c/concurrent/aeron_mpsc_rb.c
@@ -104,7 +104,7 @@ inline static int32_t aeron_mpsc_rb_claim_capacity(volatile aeron_mpsc_rb_t *rin
         aeron_rb_record_descriptor_t *record_header = (aeron_rb_record_descriptor_t *)(ring_buffer->buffer + tail_index);
 
         record_header->msg_type_id = AERON_RB_PADDING_MSG_TYPE_ID;
-        AERON_PUT_ORDERED(record_header->length, padding);
+        AERON_PUT_ORDERED(record_header->length, (int32_t)padding);
         tail_index = 0;
     }
 
@@ -133,9 +133,9 @@ aeron_rb_write_result_t aeron_mpsc_rb_write(
     {
         aeron_rb_record_descriptor_t *record_header = (aeron_rb_record_descriptor_t *)(ring_buffer->buffer + record_index);
         record_header->msg_type_id = msg_type_id;
-        AERON_PUT_ORDERED(record_header->length, -record_length);
+        AERON_PUT_ORDERED(record_header->length, -(int32_t)record_length);
         memcpy(ring_buffer->buffer + AERON_RB_MESSAGE_OFFSET(record_index), msg, length);
-        AERON_PUT_ORDERED(record_header->length, record_length);
+        AERON_PUT_ORDERED(record_header->length, (int32_t)record_length);
 
         result = AERON_RB_SUCCESS;
     }
@@ -282,7 +282,7 @@ bool aeron_mpsc_rb_unblock(volatile aeron_mpsc_rb_t *ring_buffer)
                 {
                     record = (aeron_rb_record_descriptor_t *)(ring_buffer->buffer + consumer_index);
                     record->msg_type_id = AERON_RB_PADDING_MSG_TYPE_ID;
-                    AERON_PUT_ORDERED(record->length, i - consumer_index);
+                    AERON_PUT_ORDERED(record->length, (int32_t)(i - consumer_index));
                     unblocked = true;
                 }
